requirements.txt
pyyaml
tomli
tomlkit
rich
typer
appdirs
loguru
faker
aiofiles
python-dateutil
ddddocrfix>=1.5.5
kurigram==2.2.12
tgcrypto
pillow>10.0.0
thefuzz[speedup]
aiocache
flask==3.0.3
flask-socketio==5.3.6
flask-login==0.6.3
simple-websocket
eventlet
cryptography
werkzeug==3.0.3
cachetools
emoji
numpy<2.0.0
httpx[http2]
socksio
telethon
psutil
xpinyin
curl_cffi
pymongo
pydantic
watchfiles
python-socks[asyncio]
apprise
fastapi
uvicorn[standard]
